requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pyyaml>=6.0
dataclasses-json>=0.5.7

# Scientific computing
matplotlib>=3.5.0
pandas>=1.3.0

# Configuration and serialization
pydantic>=1.8.0
jsonschema>=4.0.0

# Logging and monitoring
structlog>=22.0.0

# Optional dependencies for advanced features
# Uncomment as needed:
# openfoam-python>=0.1.0  # For CFD integration
# fenics>=2019.1.0        # For finite element analysis
# torch>=1.12.0           # For machine learning optimization